Added Baseline code
The lib folder holds the configurations for the db.

the next.config.mjs holds the dummy credentials for the database connection.

then I created some demo code in app/api/auth/[..nextauth]/route.ts.

What needs to be done now is to make to test the authentification and make modifications if the current code does not work.

We also need sessions to work as well.